Create a function to compute conditional entropy
One can compute the conditional entropy in several ways. In this
case we are providing a function that takes two distributions,
the joint distribution and the distribution of the condition.

NOTE: The entropy computed in this case is the Shannon entropy of
the conditional probability. If you want to use a different
entropy, then consider computing the conditional probability
explicity and applying the appropriate entropy function to that.